requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
click==7.1.2
colorama==0.4.4
importlib-metadata==3.4.0
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-mock==3.5.1
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.0
